split to two functions and move to helper
